feat: remove chunk storage
We use MemoryStore and providership from the kad impl now
