+++ safenode/src/client/file_apis.rs
@@ -128,7 +128,7 @@ impl Files {
     async fn upload_small(&self, small: SmallFile, verify: bool) -> Result<ChunkAddress> {
         let chunk = package_small(small)?;
         let address = *chunk.address();
-        
+
         self.client.store_chunk(chunk).await?;
 
         if verify {
