+++ safenode/src/node/mod.rs
@@ -20,10 +20,7 @@ pub use self::{
 use self::error::Error;
 
 use crate::{
-    domain::{
-        node_transfers::Transfers,
-        storage::{ChunkStorage, RegisterStorage},
-    },
+    domain::{node_transfers::Transfers, storage::RegisterStorage},
     network::Network,
 };
 
@@ -34,7 +31,6 @@ use libp2p::{Multiaddr, PeerId};
 /// storage, and broadcasts node-related events.
 pub struct Node {
     network: Network,
-    chunks: ChunkStorage,
     registers: RegisterStorage,
     transfers: Transfers,
     events_channel: NodeEventsChannel,
